Rename audiostream_rif.h to audiostream.h
This makes the naming more consistent.
